Renamed airport definition file
